requirements.txt
asgiref==3.5.2
boto3==1.26.6
botocore==1.29.6
Django==4.1.3
django-ses==3.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyotp==2.7.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
